Sketch mock L2 client
