+++ core/lib/web3_decl/Cargo.toml
@@ -21,6 +21,7 @@ pin-project-lite.workspace = true
 zksync_types.workspace = true
 zksync_config.workspace = true
 async-trait.workspace = true
+futures.workspace = true
 serde_json.workspace = true
 tokio = { workspace = true, features = ["time"] }
 tracing.workspace = true
@@ -28,7 +29,6 @@ vise.workspace = true
 
 [dev-dependencies]
 assert_matches.workspace = true
-futures.workspace = true
 rand.workspace = true
 test-casing.workspace = true
 tokio = { workspace = true, features = ["rt", "test-util"] }
